Norminette fixes
